update(api): moved utils to separate folders
