+++ api/src/utils/auth/createCookieExpiryDateInMilliseconds.ts
