--- api/src/utils/createOneTimePasscode.ts
@@ -1,4 +1,0 @@
-import crypto from 'crypto'
-export const createOneTimePasscode = () => {
-  return crypto.randomInt(100000, 999999)
-}
