+++ api/src/utils/helpers/formatSchemaErrorMessagesUtil.ts
