+++ api/src/utils/helpers/isPasswordValidUtil.ts
