+++ api/src/utils/index.ts
@@ -1,9 +1,13 @@
-export { JWTUtil } from './jwtUtil'
-export { generateHashedPassword, isPasswordCorrect } from './passwordUtil'
-export { isPasswordValid } from './isPasswordValidUtil'
-export { formatSchemaErrorMessages } from './formatSchemaErrorMessagesUtil'
-export { createCsrfToken, createHashedCsrfToken } from './csrfUtil'
-export { log } from './log'
-export { createCookieExpiryDateInMilliseconds } from './createCookieExpiryDateInMilliseconds'
-export { isEmailValid } from './isEmailValid'
-export { createOneTimePasscode } from './createOneTimePasscode'
+export { JWTUtil } from './auth/jwtUtil'
+export {
+  generateHashedPassword,
+  isPasswordCorrect,
+} from './helpers/passwordUtil'
+export { isPasswordValid } from './helpers/isPasswordValidUtil'
+export { formatSchemaErrorMessages } from './helpers/formatSchemaErrorMessagesUtil'
+export { createCsrfToken, createHashedCsrfToken } from './tokens/csrfUtil'
+export { log } from './others/log'
+export { JWTCookieUtil, CookieUtil } from './auth/cookieUtil'
+export { createCookieExpiryDateInMilliseconds } from './auth/createCookieExpiryDateInMilliseconds'
+export { isEmailValid } from './helpers/isEmailValid'
+export { createOneTimePasscode } from './tokens/createOneTimePasscode'
